feat: support ulysses sequence parallel for FSDP (#278)
* PullRequest: 686 [FEAT] support SGLang dp-attention and cross-nodes with slurm

Merge branch sxj/cross-nodes-sglang of git@code.alipay.com:inclusionAI/AReaL.git into gh
https://code.alipay.com/inclusionAI/AReaL/pull_requests/686



* PullRequest: 639 [FEAT][EXPERIMENTAL] add megatron dcp
* PullRequest: 654 opt: optimizing megatron model loading from HF checkpoints
* [FEAT] support SGLang dp-attention and cross-nodes
* format
* fix
* fix
* fix slurm
* .
* del sep_env_vars
* fix name
* fix

* PullRequest: 695 fix qwen3 moe attention_mask

Merge branch sxj/fix-moe of git@code.alipay.com:inclusionAI/AReaL.git into gh
https://code.alipay.com/inclusionAI/AReaL/pull_requests/695



* fix moe
* format

* PullRequest: 696 fix eval import

Merge branch sxj/fix-eval-import of git@code.alipay.com:inclusionAI/AReaL.git into gh
https://code.alipay.com/inclusionAI/AReaL/pull_requests/696



* fix import

* PullRequest: 698 [Feature] [Experimental] Add Megatron SFT examples

Merge branch mzy/megatron-train of git@code.alipay.com:inclusionAI/AReaL.git into gh
https://code.alipay.com/inclusionAI/AReaL/pull_requests/698?tab=diff



* PullRequest: 605 [Fix] FIx a bug in slurm launcher.
* PullRequest: 608 [FIX] fix slurm and ray
* experimental megatron engine init
* add simple forward and train for testing
* run
* try mbridge
* .
* save load test
* Cherry-pick commit 63d24e1f "PullRequest: 601 [FI..." 到当前分支
* Cherry-pick commit 8afd706b "PullRequest: 596 [FI..." 到当前分支
* PullRequest: 612 [Feature] [Experimental] Add model initialization for MegatronEngine
* PullRequest: 616 [Feature] [Experimental] Add single GPU train and forward in MegatronEngine for testing purposes.
* .
* .
* add 5d parallel allocation mode
* allocation mode
* PullRequest: 617 [Feature] [Experimental] Add Megatron model save/load in HF format with mbridge.
* .
* pass tests
* forward
* add one distributed test
* .
* debug broadcast
* .
* .
* .
* .
* tp debug
* tp gather result
* .
* compare forward result
* tp+pp tested
* .
* .
* .
* .
* merge
* refactor padding
* .
* fix
* .
* .
* change max_tokens_per_mb to max_tokens_per_gpu
* .
* fix bug
* PullRequest: 639 [FEAT][EXPERIMENTAL] add megatron dcp
* PullRequest: 654 opt: optimizing megatron model loading from HF checkpoints
* fix allocation mode
* megatron tests
* before test
* .
* .
* remove allocation mode in cli_args.py
* add train_batch
* .
* .
* .
* .
* fix loss
* .
* add configs
* before fixing stats
* add broadcast stats
* broadcast stats_tracker
* fix bcast
* .
* clear
* .
* .
* separate create_process_groups from initialize
* .
* .
* .
* .

* PullRequest: 699 fix slurm random-seed same in one node

Merge branch sxj/fix-slurm-random-seed of git@code.alipay.com:inclusionAI/AReaL.git into gh
https://code.alipay.com/inclusionAI/AReaL/pull_requests/699



* fix slurm random-seed same in one node

* PullRequest: 704 [FEAT] support SGLang dp-attention and cross-nodes with ray

Merge branch sxj/cross-nodes-sglang-ray of git@code.alipay.com:inclusionAI/AReaL.git into gh
https://code.alipay.com/inclusionAI/AReaL/pull_requests/704?tab=diff



* [FEAT] support SGLang dp-attention and cross-nodes with ray
* fix env_vars
* .

* PullRequest: 721 [FIX] fix random seed

Merge branch sxj/fix-random-seed of git@code.alipay.com:inclusionAI/AReaL.git into gh
https://code.alipay.com/inclusionAI/AReaL/pull_requests/721



* [FIX] fix random seed

* PullRequest: 722 [FEAT] add sglang ep_size config

Merge branch sxj/add-sglang-ep of git@code.alipay.com:inclusionAI/AReaL.git into gh
https://code.alipay.com/inclusionAI/AReaL/pull_requests/722



* [FEAT] add sglang ep_size config

* PullRequest: 706 Support DeepSpeed-Ulysses (sequence parallelism)

Merge branch zwt/ulysses of git@code.alipay.com:inclusionAI/AReaL.git into gh
https://code.alipay.com/inclusionAI/AReaL/pull_requests/706



* Support DeepSpeed-Ulysses (sequence parallelism)

* .

---------

+++ areal/tests/test_utils.py
@@ -44,7 +44,9 @@ def test_micro_batch_split(mock_padded_data, n_mbs, max_tokens_per_gpu):
     # Unpad and split to microbatches
     packed_data = pack_tensor_dict(mock_padded_data)
     original_lens = packed_data["cu_seqlens"][1:] - packed_data["cu_seqlens"][:-1]
-    assert torch.allclose(original_lens, mock_padded_data["attention_mask"].sum(1))
+    assert torch.allclose(
+        original_lens.long(), mock_padded_data["attention_mask"].sum(1)
+    )
     split_result = split_padded_tensor_dict_into_mb_list(mock_padded_data, mb_spec)
     split_result.mbs = [pack_tensor_dict(mb) for mb in split_result.mbs]
     reordered_lens = [original_lens[i] for i in split_result.forward_indices]
